:lipstick: style: font sizes changed and noise background converted from png to webp
