+++ components/svg/index.ts
@@ -1,4 +1,4 @@
 export { default as ArrowRightSVG } from './arrow-right';
 export { default as BarsSVG } from './bars';
 export { default as LogoSVG } from './logo';
-export { default as MoveBitSVG } from './moveBit';
+export { default as MoveBitSVG } from './move-bit';
